Extract Sidebar into standalone Svelte component
Moves sidebar state (sidebarCollapsed, expandedGroups, menuProfile), derived
grouped, all sidebar functions (toggleSidebar, toggleGroup, isGroupExpanded,
groupByProject, relativeTime), sidebar template, and sidebar CSS into
src/lib/components/Sidebar.svelte. +page.svelte uses the new component via
bind:this to expose toggleSidebar for the Cmd+B keyboard shortcut.
