+++ src/routes/+page.svelte
@@ -20,6 +20,7 @@
   import UsageDashboard from "$lib/components/UsageDashboard.svelte";
   import NewSessionModal from "$lib/components/NewSessionModal.svelte";
   import { getPurposePrompt } from "$lib/prompts";
+  import Sidebar from "$lib/components/Sidebar.svelte";
 
   let profiles = $state([]);
   let activeProfile = $state(null);
@@ -33,44 +34,10 @@
   let tokenInterval = null;
   let appVersion = $state('');
   let claudePlan = $state('');
-  let sidebarCollapsed = $state(
-    typeof localStorage !== 'undefined' ? localStorage.getItem('clauge-sidebar-collapsed') === 'true' : false
-  );
-
-  function toggleSidebar() {
-    sidebarCollapsed = !sidebarCollapsed;
-    localStorage.setItem('clauge-sidebar-collapsed', String(sidebarCollapsed));
-    // Refit all terminals after transition
-    setTimeout(() => {
-      for (const [, entry] of terminalStore.terminalMap) {
-        if (entry.fitAddon && entry.container.offsetWidth > 0) {
-          try { entry.fitAddon.fit(); } catch(_) {}
-        }
-      }
-    }, 250);
-  }
-
-  // Expand/collapse state — persisted to localStorage
-  let expandedGroups = $state(
-    typeof localStorage !== 'undefined'
-      ? JSON.parse(localStorage.getItem('clauge-expanded') || '{}')
-      : {}
-  );
-
-  function toggleGroup(name) {
-    expandedGroups[name] = !expandedGroups[name];
-    expandedGroups = { ...expandedGroups }; // trigger reactivity
-    localStorage.setItem('clauge-expanded', JSON.stringify(expandedGroups));
-  }
-
-  function isGroupExpanded(name) {
-    // Default to expanded if not set
-    return expandedGroups[name] !== false;
-  }
+  let sidebarRef;
 
   // Delete confirmation
   let deleteConfirm = $state(null); // profile to confirm delete
-  let menuProfile = $state(null); // profile whose ellipsis menu is open
   let profileMenuOpen = $state(false);
   // Git state lives in gitStore
 
@@ -547,8 +514,8 @@
   }
 
   function deleteProfile(e, profile) {
-    e.preventDefault();
-    e.stopPropagation();
+    e?.preventDefault();
+    e?.stopPropagation();
     deleteConfirm = profile;
   }
 
@@ -601,27 +568,6 @@
     await loadProfiles();
   }
 
-  function relativeTime(iso) {
-    if (!iso) return "";
-    const sec = Math.floor((Date.now() - new Date(iso).getTime()) / 1000);
-    if (sec < 60) return "just now";
-    if (sec < 3600) return Math.floor(sec/60) + "m ago";
-    if (sec < 86400) return Math.floor(sec/3600) + "h ago";
-    return Math.floor(sec/86400) + "d ago";
-  }
-
-  function groupByProject(list) {
-    const g = {};
-    for (const p of list) {
-      const name = p.projectName || "Unknown";
-      if (!g[name]) g[name] = [];
-      g[name].push(p);
-    }
-    return g;
-  }
-
-  let grouped = $derived(groupByProject(profiles));
-
   // Git helpers — delegates to gitStore, passing resolved projectPath
   function gitPath() { return activeProfile?.worktreePath || activeProfile?.projectPath || ''; }
   function gitProfileId() { return activeProfile?.id; }
@@ -670,7 +616,7 @@
       const idx = parseInt(e.key) - 1;
       if (profiles[idx]) selectProfile(profiles[idx]);
     }
-    if (e.metaKey && e.key === 'b') { e.preventDefault(); toggleSidebar(); }
+    if (e.metaKey && e.key === 'b') { e.preventDefault(); sidebarRef?.toggleSidebar(); }
     if (e.metaKey && e.key === 'l') { e.preventDefault(); toggleShell(); }
     if (e.key === 'Escape') { showModal = false; showSettings = false; }
   }
@@ -758,85 +704,22 @@
   });
 </script>
 
-<svelte:window onkeydown={handleGlobalKeydown} onresize={handleWindowResize} onclick={() => { menuProfile = null; profileMenuOpen = false; gitStore.gitPanelOpen = false; contextsStore.showContextDropdown = false; }} oncontextmenu={(e) => { if (!import.meta.env.DEV) e.preventDefault(); }} />
+<svelte:window onkeydown={handleGlobalKeydown} onresize={handleWindowResize} onclick={() => { profileMenuOpen = false; gitStore.gitPanelOpen = false; contextsStore.showContextDropdown = false; }} oncontextmenu={(e) => { if (!import.meta.env.DEV) e.preventDefault(); }} />
 
 <div class="app-wrapper">
 <div class="app">
   <!-- svelte-ignore a11y_no_static_element_interactions -->
   <div class="drag-bar" onmousedown={handleDragStart}></div>
-  <aside class="sidebar" class:collapsed={sidebarCollapsed}>
-    <div class="sidebar-header">
-      <span class="app-title">Clauge {#if claudePlan}<span class="plan-badge">{claudePlan}</span>{/if}</span>
-      <div class="header-actions">
-        <button class="new-btn" onclick={() => { showModal = true; contextsStore.loadContextSnippets(); }} title="New Session (Cmd+N)">+</button>
-      </div>
-    </div>
-    <div class="sidebar-content">
-      {#if profiles.length === 0}
-        <div class="empty-sidebar">No sessions yet. Click + to create one.</div>
-      {:else}
-        {#each Object.entries(grouped) as [projectName, items]}
-          <div class="project-group">
-            <button class="project-header" onclick={() => toggleGroup(projectName)}>
-              <svg class="chevron" class:collapsed={!isGroupExpanded(projectName)} width="10" height="10" viewBox="0 0 10 10" fill="currentColor">
-                <path d="M3 2l4 3-4 3z"/>
-              </svg>
-              {projectName}
-              <span class="project-count">{items.length}</span>
-            </button>
-            {#if isGroupExpanded(projectName)}
-              {#each items as profile}
-                  <button
-                    class="profile-item"
-                    class:active={activeProfile?.id === profile.id}
-                    onclick={() => selectProfile(profile)}
-                  >
-                    <div class="profile-title">
-                      <span class="status-dot" class:active={activeProfile?.id === profile.id} class:bg-active={terminalStore.sessionActivity[profile.id] === 'active'} class:bg-done={terminalStore.sessionActivity[profile.id] === 'done'}></span>
-                      {profile.title}
-                    </div>
-                    <div class="profile-meta">
-                      <span class="badge" style="color:{purposeColors[profile.purpose] || '#8b949e'}; background:{purposeColors[profile.purpose] || '#8b949e'}22">{profile.purpose}</span>
-                      {#if profile.worktreeBranch}
-                        <span class="wt-badge" title="Isolated worktree: {profile.worktreeBranch}">WT</span>
-                      {/if}
-                      <span class="time">{relativeTime(profile.lastUsedAt)}</span>
-                    </div>
-                    <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-                    <span class="ellipsis-btn" onclick={(e) => { e.stopPropagation(); menuProfile = menuProfile?.id === profile.id ? null : profile; }}>
-                      <svg width="14" height="14" viewBox="0 0 16 16" fill="currentColor"><circle cx="8" cy="3" r="1.5"/><circle cx="8" cy="8" r="1.5"/><circle cx="8" cy="13" r="1.5"/></svg>
-                    </span>
-                    {#if menuProfile?.id === profile.id}
-                      <!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
-                      <div class="session-menu" onclick={(e) => e.stopPropagation()}>
-                        <button class="session-menu-item" onclick={() => { menuProfile = null; contextsStore.showContextPicker = profile; contextsStore.loadContextSnippets(); }}>
-                          <svg width="12" height="12" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round"><path d="M14 2H6a2 2 0 00-2 2v16a2 2 0 002 2h12a2 2 0 002-2V8z"/><polyline points="14 2 14 8 20 8"/></svg>
-                          Add Contexts
-                        </button>
-                        <button class="session-menu-item danger" onclick={() => { menuProfile = null; deleteConfirm = profile; }}>
-                          <svg width="12" height="12" viewBox="0 0 16 16" fill="currentColor"><path d="M6.5 1.75a.25.25 0 01.25-.25h2.5a.25.25 0 01.25.25V3h-3V1.75zM11 3V1.75A1.75 1.75 0 009.25 0h-2.5A1.75 1.75 0 005 1.75V3H2.75a.75.75 0 000 1.5h.928l.747 10.218A1.75 1.75 0 006.172 16h3.656a1.75 1.75 0 001.747-1.282L12.322 4.5h.928a.75.75 0 000-1.5H11zm-5.522 1.5l.735 10.06a.25.25 0 00.249.19h3.076a.25.25 0 00.249-.19l.735-10.06H5.478z"/></svg>
-                          Delete
-                        </button>
-                      </div>
-                    {/if}
-                  </button>
-              {/each}
-            {/if}
-          </div>
-        {/each}
-      {/if}
-    </div>
-  </aside>
-
-  <button class="sidebar-toggle" onclick={toggleSidebar} title={sidebarCollapsed ? 'Expand sidebar (Cmd+B)' : 'Collapse sidebar (Cmd+B)'}>
-    <svg width="12" height="12" viewBox="0 0 12 12" fill="currentColor">
-      {#if sidebarCollapsed}
-        <path d="M4 1l5 5-5 5z"/>
-      {:else}
-        <path d="M8 1L3 6l5 5z"/>
-      {/if}
-    </svg>
-  </button>
+  <Sidebar
+    bind:this={sidebarRef}
+    {profiles}
+    {activeProfile}
+    {claudePlan}
+    onSelect={selectProfile}
+    onNewSession={() => showModal = true}
+    onDeleteProfile={deleteProfile}
+    onRefitTerminals={handleWindowResize}
+  />
 
   <div class="terminal-wrapper" bind:this={wrapperEl}>
     <!-- svelte-ignore a11y_no_static_element_interactions -->
@@ -966,17 +849,6 @@
   .app-wrapper { display: flex; flex-direction: column; height: 100vh; width: 100vw; overflow: hidden; }
   .app { display: flex; flex: 1; min-height: 0; overflow: hidden; background: transparent; }
 
-  .sidebar { width: 220px; min-width: 220px; background: var(--sidebar-bg); border-right: 1px solid var(--border); display: flex; flex-direction: column; user-select: none; transition: width 0.2s ease, min-width 0.2s ease, opacity 0.2s ease; overflow: hidden; }
-  .sidebar.collapsed { width: 0; min-width: 0; border-right: none; opacity: 0; pointer-events: none; }
-  .sidebar-toggle { position: absolute; left: 220px; top: 50%; transform: translateY(-50%); z-index: 50; width: 12px; height: 28px; border: none; border-radius: 0 4px 4px 0; background: transparent; color: var(--text-secondary); cursor: pointer; display: flex; align-items: center; justify-content: center; transition: left 0.2s ease, background 0.15s, opacity 0.15s; -webkit-app-region: no-drag; opacity: 0; }
-  .sidebar-toggle:hover, .app:hover .sidebar-toggle { opacity: 1; background: rgba(255,255,255,0.06); color: var(--text-primary); }
-  .sidebar.collapsed ~ .sidebar-toggle { left: 0; }
-  .sidebar-header { display: flex; align-items: center; justify-content: space-between; padding: 14px; padding-top: 38px; border-bottom: 1px solid var(--border); }
-  .app-title { font-size: 15px; font-weight: 700; color: var(--text-primary); display: flex; align-items: center; gap: 6px; }
-  .plan-badge { font-size: 9px; font-weight: 700; text-transform: uppercase; letter-spacing: 0.8px; padding: 2px 6px; border-radius: 4px; background: linear-gradient(135deg, rgba(255,215,0,0.15), rgba(255,170,50,0.1)); color: #ffd700; border: 1px solid rgba(255,215,0,0.3); position: relative; overflow: hidden; }
-  .plan-badge::after { content: ''; position: absolute; top: -50%; left: -100%; width: 60%; height: 200%; background: linear-gradient(90deg, transparent, rgba(255,215,0,0.2), transparent); animation: shine 3s ease-in-out infinite; }
-  @keyframes shine { 0% { left: -100%; } 50% { left: 150%; } 100% { left: 150%; } }
-  .header-actions { display: flex; gap: 6px; align-items: center; -webkit-app-region: no-drag; }
   .profile-wrap { position: relative; }
   .profile-avatar { width: 22px; height: 22px; border-radius: 50%; border: none; background: linear-gradient(135deg, var(--accent), color-mix(in srgb, var(--accent) 60%, #000)); color: #fff; cursor: pointer; display: flex; align-items: center; justify-content: center; transition: opacity 0.15s; padding: 0; overflow: hidden; }
   .profile-avatar:hover { opacity: 0.85; }
@@ -991,43 +863,6 @@
   .pm-coffee:hover { color: #e3b341; }
   .pm-coffee:hover svg { stroke: #e3b341; }
   .pm-external { width: 11px !important; height: 11px !important; margin-left: auto; opacity: 0.4; }
-  .new-btn { width: 28px; height: 28px; border-radius: 6px; border: 1px solid var(--border); background: transparent; color: var(--text-primary); font-size: 18px; cursor: pointer; display: flex; align-items: center; justify-content: center; -webkit-app-region: no-drag; transition: all 0.2s cubic-bezier(0.34, 1.56, 0.64, 1); }
-  .new-btn:hover { background: var(--border); transform: scale(1.1); }
-  .new-btn:active { transform: scale(0.95); }
-
-  .sidebar-content { flex: 1; overflow-y: auto; padding: 8px 0; -webkit-app-region: no-drag; }
-  .empty-sidebar { padding: 24px 14px; text-align: center; color: var(--text-secondary); font-size: 13px; }
-  .project-group { margin-bottom: 2px; }
-  .project-header { display: flex; align-items: center; gap: 4px; width: 100%; padding: 6px 14px; font-size: 11px; font-weight: 600; color: var(--text-secondary); text-transform: uppercase; letter-spacing: 0.5px; border: none; background: transparent; cursor: pointer; font-family: inherit; transition: color 0.15s; }
-  .project-header:hover { color: var(--text-primary); }
-  .chevron { transition: transform 0.15s; flex-shrink: 0; }
-  .chevron.collapsed { transform: rotate(0deg); }
-  .chevron:not(.collapsed) { transform: rotate(90deg); }
-  .project-count { margin-left: auto; font-size: 10px; color: var(--text-secondary); opacity: 0.6; font-weight: 400; }
-  .delete-confirm { padding: 8px 14px; font-size: 12px; color: var(--text-primary); display: flex; flex-direction: column; gap: 6px; animation: fadeIn 0.15s ease-out; }
-  .delete-actions { display: flex; gap: 6px; }
-  .del-yes { padding: 3px 10px; border-radius: 4px; border: 1px solid #f85149; background: transparent; color: #f85149; font-size: 11px; cursor: pointer; font-family: inherit; transition: all 0.15s; }
-  .del-yes:hover { background: #f85149; color: #fff; }
-  .del-no { padding: 3px 10px; border-radius: 4px; border: 1px solid var(--border); background: transparent; color: var(--text-secondary); font-size: 11px; cursor: pointer; font-family: inherit; }
-  .del-no:hover { color: var(--text-primary); }
-  .profile-item { width: 100%; display: block; text-align: left; padding: 8px 14px; border: none; background: transparent; cursor: pointer; border-left: 3px solid transparent; font-family: inherit; -webkit-app-region: no-drag; position: relative; }
-  .profile-item:hover { background: var(--hover-bg, rgba(255,255,255,0.06)); }
-  .profile-item.active { background: rgba(31,111,235,0.15); border-left-color: var(--accent); box-shadow: inset 0 0 20px rgba(88, 166, 255, 0.08); }
-  .profile-title { font-size: 13px; font-weight: 500; color: var(--text-primary); margin-bottom: 3px; }
-  .profile-meta { display: flex; align-items: center; justify-content: space-between; }
-  .badge { font-size: 10px; font-weight: 600; padding: 1px 6px; border-radius: 10px; }
-  .wt-badge { font-size: 8px; font-weight: 700; padding: 1px 4px; border-radius: 3px; background: rgba(210, 168, 255, 0.2); color: #d2a8ff; letter-spacing: 0.5px; }
-
-  .profile-item { padding-right: 28px; }
-  .ellipsis-btn { position: absolute; right: 6px; top: 50%; transform: translateY(-50%); opacity: 0; padding: 4px; border-radius: 4px; color: var(--text-secondary); cursor: pointer; transition: opacity 0.15s, background 0.15s; z-index: 2; }
-  .profile-item:hover .ellipsis-btn { opacity: 1; }
-  .ellipsis-btn:hover { background: var(--hover-bg, rgba(255,255,255,0.08)); color: var(--text-primary); }
-
-  .session-menu { position: absolute; right: 6px; top: calc(50% + 14px); background: #1c2128; border: 1px solid var(--border); border-radius: 8px; padding: 4px; min-width: 110px; box-shadow: 0 8px 24px rgba(0,0,0,0.4); z-index: 10; animation: fadeIn 0.1s ease-out; }
-  .session-menu-item { display: flex; align-items: center; gap: 6px; width: 100%; padding: 6px 10px; border: none; background: transparent; color: var(--text-secondary); font-size: 12px; font-family: inherit; cursor: pointer; border-radius: 5px; transition: background 0.12s; }
-  .session-menu-item:hover { background: rgba(255,255,255,0.06); }
-  .session-menu-item.danger:hover { background: rgba(248,81,73,0.12); color: #f85149; }
-  .time { font-size: 11px; color: var(--text-secondary); }
   .bottom-bar { display: flex; align-items: center; padding: 3px 16px; background: var(--sidebar-bg); border-top: 1px solid var(--border); flex-shrink: 0; position: relative; }
   .bottom-left { display: flex; align-items: center; gap: 8px; flex-shrink: 0; }
   .bottom-center { position: absolute; left: 50%; transform: translateX(-50%); display: flex; align-items: center; justify-content: center; gap: 12px; }
@@ -1261,11 +1096,6 @@
   .modal-actions { display: flex; justify-content: flex-end; gap: 8px; margin-top: 16px; }
   .modal-actions button { padding: 7px 16px; border-radius: 6px; font-size: 13px; cursor: pointer; border: 1px solid var(--border); background: var(--btn-bg, #21262d); color: var(--text-primary); font-family: inherit; }
 
-  .status-dot { display: inline-block; width: 6px; height: 6px; border-radius: 50%; background: #484f58; margin-right: 6px; vertical-align: middle; transition: background 0.3s; }
-  .status-dot.active { background: #3fb950; box-shadow: 0 0 6px rgba(63, 185, 80, 0.5); }
-  .status-dot.bg-active { background: var(--accent); box-shadow: 0 0 6px color-mix(in srgb, var(--accent) 50%, transparent); animation: bgPulse 0.8s ease-in-out infinite; }
-  .status-dot.bg-done { background: #d29922; box-shadow: 0 0 6px rgba(210, 153, 34, 0.5); }
-  @keyframes bgPulse { 0%, 100% { opacity: 1; transform: scale(1); } 50% { opacity: 0.4; transform: scale(0.7); } }
   @keyframes pulse { 0%, 100% { opacity: 1; } 50% { opacity: 0.5; } }
   .purpose-glow { position: absolute; top: 0; left: 0; right: 0; height: 60px; z-index: 1; pointer-events: none; animation: glowFadeIn 0.5s ease-out; }
   @keyframes glowFadeIn { from { opacity: 0; } to { opacity: 1; } }
